Initializing gameplay panel
Now, in this commit, we are able to set, initialize, and prepare the gameplay panel in which it contains the supposed background image, "grass_biome-BG_lv1.png", which can be found on the folder "assets/backgrounds/". 
Moreover,  there is an update to the core files as well as some code cleanup which includes implementing more listeners particularly ActionListener and MouseListener. I also made sure that at least nothing leaks, and I hope everything will run well!