Created a script to clean bin
In this commit, I have created a new script file with a .sh extension, so, therefore, it is a unix-based commands.
This files immediately cleans all containment of bin in which it removes all *.class files. 
Please proceed with caution as this may cause compilation problems! This is only applied when there are problems running and compiling with java 

**It is highly advised to recompile again using 'compile.sh' found in src**