Trying to render apple image, but didn't work!
In this commit, I try to render the apple image over the JPanel that is extending our class, however, I was not able to do so because I cannot render the image over a specific point x coordinate and y coordinate. Additionally, I have turned off the drawing of gridlines as they seem to be distracting. Maybe we can turn this into a feature!