Fixed Home Window buttons issue
In this commit, I have fixed the occurring issue about the hovering and boundaries of buttons found in home window. Therefore, the fix I have applied is repositioning and retexturizing the Play and About buttons. Moreover, we have also updated the list of actions and responses to these buttons such as the action event and mouse event. Lastly, we have also improved the boundaries and setting of each icon so that nothing can overlap!