Various Improvements and Update to Snake
In this commit, we have finally improved some features and fixes for the Snake, that is to say, we have applied a new implementation system to render our own snake character that elongates as it eats an apple. Moreover, we still need more patches and fixes to improve the quality of this new feature. 

Updates and fixes include: 
- Rendering an elongating Snake based off a graphical texture found in 'assets/snake' directory 
- Implemented a panel to which the Snake will be instantiated 
- Added movement, collision detection, and apple coordinates systems 
- Added sound upon snake movement

Issues: 
- Sometimes the apple disappears out of the screen, this may be caused by the apple x and y coordinates 
- The snake character textures do not rotate accordingly relative to its direction and movement. IMPLEMENT A TEXTURE ROTATION FUNCTION!