Various patches and improvements to Snake
Fix and patches applied to the Snake's head and body chunk rendering and positioning according to its direction and movement. Moreover, we have removed the tail so as to make things much easier. Along with this, we have also applied a new texture for the Snake character with a much crispier and better design as well as proper sizing and rendering of graphical objects.

Patches and improvements:
- Fixed lining up and position of head and body
- Improved Snake texture for proper placement on the graphical methods and such

Need improvements or issues:
- The rest of the body and head do not rotate in relative to where it is facing